Add misago.categories.admin and misago.legal.admin packages
