+++ misago/categories/admin/forms.py
@@ -4,11 +4,11 @@ from django.utils.html import conditional_escape, mark_safe
 from django.utils.translation import gettext_lazy as _
 from mptt.forms import TreeNodeChoiceField, TreeNodeMultipleChoiceField
 
-from . import THREADS_ROOT_NAME
-from ..admin.forms import YesNoSwitch
-from ..core.validators import validate_sluggable
-from ..threads.threadtypes import trees_map
-from .models import Category, CategoryRole
+from ...admin.forms import YesNoSwitch
+from ...core.validators import validate_sluggable
+from ...threads.threadtypes import trees_map
+from .. import THREADS_ROOT_NAME
+from ..models import Category, CategoryRole
 
 
 class AdminCategoryFieldMixin:
