+++ misago/categories/admin/tests/test_permissions_admin_views.py
@@ -1,11 +1,11 @@
 from django.urls import reverse
 
-from ...acl import ACL_CACHE
-from ...acl.models import Role
-from ...acl.admin.test import mock_role_form_data
-from ...admin.test import AdminTestCase
-from ...cache.test import assert_invalidates_cache
-from ..models import Category, CategoryRole
+from ....acl import ACL_CACHE
+from ....acl.models import Role
+from ....acl.admin.test import mock_role_form_data
+from ....admin.test import AdminTestCase
+from ....cache.test import assert_invalidates_cache
+from ...models import Category, CategoryRole
 
 
 def create_data(data_dict):
