+++ misago/categories/admin/views/perms.py
@@ -2,18 +2,18 @@ from django.contrib import messages
 from django.shortcuts import redirect
 from django.utils.translation import gettext_lazy as _
 
-from ...acl.admin.forms import get_permissions_forms
-from ...acl.admin.views import RoleAdmin, RolesList
-from ...acl.cache import clear_acl_cache
-from ...acl.models import Role
-from ...admin.views import generic
+from ....acl.admin.forms import get_permissions_forms
+from ....acl.admin.views import RoleAdmin, RolesList
+from ....acl.cache import clear_acl_cache
+from ....acl.models import Role
+from ....admin.views import generic
+from ...models import Category, CategoryRole, RoleCategoryACL
 from ..forms import (
     CategoryRoleForm,
     CategoryRolesACLFormFactory,
     RoleCategoryACLFormFactory,
 )
-from ..models import Category, CategoryRole, RoleCategoryACL
-from .categoriesadmin import CategoriesList, CategoryAdmin
+from .categories import CategoriesList, CategoryAdmin
 
 
 class CategoryRoleAdmin(generic.AdminBaseMixin):
