+++ misago/legal/utils.py
@@ -8,24 +8,6 @@ from ..markup import common_flavour
 from .models import Agreement, UserAgreement
 
 
-def set_agreement_as_active(agreement, commit=False):
-    agreement.is_active = True
-    Agreement.objects.filter(type=agreement.type).exclude(pk=agreement.pk).update(
-        is_active=False
-    )
-
-    if commit:
-        agreement.save(update_fields=["is_active"])
-        Agreement.objects.invalidate_cache()
-
-
-def disable_agreement(agreement, commit=False):
-    agreement.is_active = False
-    if commit:
-        agreement.save(update_fields=["is_active"])
-        Agreement.objects.invalidate_cache()
-
-
 def get_required_user_agreement(user, agreements):
     if user.is_anonymous:
         return None
