--- misago/legal/views/__init__.py
@@ -1,1 +1,0 @@
-from .legal import privacy_policy, terms_of_service
